requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.25.2
websockets==12.0
pika==1.3.2
redis==5.0.1
pytesseract==0.3.10
boto3==1.34.0
python-dotenv==1.0.0
numpy==1.24.3
PyJWT==2.8.0
cryptography==41.0.8
pydantic==2.5.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.17.0
Pillow==10.1.0
python-multipart==0.0.6
fastapi-poe==0.1.0
openai==1.3.0
